// src/app/api/models/event.interface.ts
// https://api.dev.2ticket.pt/events/apikey/d1b53c7d-22c4-4b75-92dd-c8966317fbba/list/public
import { EventTicketType } from './event-ticket-type.interface';
import { Participant } from './participant.interface';
import { Place } from './place.interface';

export interface Event {
  id: number;
  name: string;
  name_es: string | null;
  name_en: string | null;
  slug: string;
  season: string | null;
  number: string | null;
  eventOpeningTime: string | null;
  eventClosingTime: string | null;
  outOfHoursBehaviour: string;
  eventTime: string;
  eventQuotaExpiration: string;
  eventQuotaValidation: string;
  eventTypeId: number;
  purchaseLimitAssociate: number;
  purchaseLimitRegular: number;
  videoFrame: string | null;
  ageRating: string | null;
  duration: string | null;
  local: string | null;
  listThumbnail?: string;
  description: string | null;
  description_es: string | null;
  description_en: string | null;
  promoter: string | null;
  notes: string | null;
  notes_es: string | null;
  notes_en: string | null;
  externalURL: string | null;
  purchaseType: string;
  eventTypefilename: string;
  places: Place[];
  participants: Participant[];
  ticketTypes: EventTicketType[];
  ready: boolean;
  readyForInvites: boolean;
  active: boolean;
  activeOnline: boolean;
  requirePaymentParticipant: boolean;
  requiresId: boolean;
  walletCodeVisible: boolean;
  invoiceUnlockedDatetime: string | null;
  idLiga: string | null;
  recurringPeriods: string[];
}